use different docker image names for server and client to add flexibility in makefile; support supplying docker image build flags

+++ Makefile
@@ -87,19 +87,24 @@ build:
 # docker targets
 
 DOCKER_IMAGE_SOURCE_BRANCH ?=  # empty means copying the whole current repo
+DOCKER_BUILD_FLAGS ?=
 DOCKER_SERVER_FLAGS ?= --net=host -d
 DOCKER_CLIENT_FLAGS ?= --net=host --rm
 SERVER_RESTART ?= always
 SERVER_PORT ?= 16099
 CLIENT_ENV ?=
 
-DOCKER_IMAGE_NAME ?= chat-terminal:latest
+DOCKER_SERVER_IMAGE_NAME ?= chat-terminal-server
+DOCKER_CLIENT_IMAGE_NAME  ?= chat-terminal-client
 DOCKER_SERVER_CONTAINER_NAME ?= chat-terminal-server
 DOCKER_CLIENT_CONTAINER_NAME ?= chat-terminal-client
 
 .PHONY: docker-build-image docker-setup docker-run-server docker-run-client
 
 docker-build-image:
+	if [ -z "$(DOCKER_IMAGE_NAME)" ]; then \
+		echo "Please specify DOCKER_IMAGE_NAME"; exit 1; \
+	fi
 	@echo $(DOCKER_PHRASE) "Building docker image..."
 	docker image rm -f $(DOCKER_IMAGE_NAME) 2>/dev/null || true
 ifneq ($(DOCKER_IMAGE_SOURCE_BRANCH),)
@@ -107,17 +112,18 @@ ifneq ($(DOCKER_IMAGE_SOURCE_BRANCH),)
 	mkdir -p ./tmp
 	tmp_git_archive=$$(mktemp ./tmp/chat-terminal-repo-archive-XXXXXX.tar.gz) &&	\
 		git archive --format=tar.gz -o $$tmp_git_archive master && \
-		docker build -t $(DOCKER_IMAGE_NAME) --build-arg REPO_ARCHIVE=$$tmp_git_archive . && \
+		docker build -t $(DOCKER_IMAGE_NAME) --build-arg REPO_ARCHIVE=$$tmp_git_archive $(DOCKER_BUILD_FLAGS) . && \
 		rm $$tmp_git_archive
 	rm -d ./tmp 2>/dev/null || true
 else
 # build from the current repo
-	docker build -t $(DOCKER_IMAGE_NAME) .
+	docker build -t $(DOCKER_IMAGE_NAME) $(DOCKER_BUILD_FLAGS) .
 endif
 
 docker-setup: docker-build-image
 	@echo $(DOCKER_PHRASE) "Running setup in docker..."
 
+docker-run-server: DOCKER_IMAGE_NAME = $(DOCKER_SERVER_IMAGE_NAME):latest
 docker-run-server: docker-build-image
 	@echo $(DOCKER_PHRASE) "Running server in docker..."
 	if $$(docker ps -a -q --filter "name=$(DOCKER_SERVER_CONTAINER_NAME)" | grep -q .); then \
@@ -126,6 +132,7 @@ docker-run-server: docker-build-image
 	fi
 	docker run --name $(DOCKER_SERVER_CONTAINER_NAME) $(DOCKER_SERVER_FLAGS) --restart $(SERVER_RESTART) -p $(SERVER_PORT):$(SERVER_PORT) $(DOCKER_IMAGE_NAME) --host 0.0.0.0 --port $(SERVER_PORT)
 
+docker-run-client: DOCKER_IMAGE_NAME = $(DOCKER_CLIENT_IMAGE_NAME):latest
 docker-run-client: docker-build-image
 	@echo $(DOCKER_PHRASE) "Running client in docker..."
 	docker run --name $(DOCKER_CLIENT_CONTAINER_NAME) -it $(DOCKER_CLIENT_FLAGS) --entrypoint /bin/bash $(DOCKER_IMAGE_NAME) -c 'source ~/.chat-terminal/chat-terminal.sh && $(CLIENT_ENV) chat-terminal'
